Fix cycle between configuration and logging
